moved wine-search filter
