+++ my-app/src/app/model/wine-filter.ts
